fix tests bug

+++ controller/src/main/java/controllerLogic/CoreManager.java
@@ -14,6 +14,7 @@ import storage.ReportManager;
 import utils.VectorFind;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Vector;
 
@@ -85,11 +86,20 @@ public class CoreManager {
 
         try {
             FileUtils.deleteDirectory(new File(System.getProperty("user.home") + "\\.temevi" + "\\htmlCoverage"));
-            FileUtils.forceDelete(new File(System.getProperty("user.home") + "\\.temevi" + "\\coverage.csv"));
+            File covergeFile=new File(System.getProperty("user.home") + "\\.temevi" + "\\coverage.csv");
+            if(covergeFile.exists()){
+                FileUtils.forceDelete(new File(System.getProperty("user.home") + "\\.temevi" + "\\coverage.csv"));
+            }
             FileUtils.forceDelete(new File(System.getProperty("user.home") + "\\.temevi" + "\\jacoco.exec"));
-            //TODO vedere perchè non cancella pitreport
-            FileUtils.deleteDirectory(new File(System.getProperty("user.home") + "\\.temevi\\pitreport"));
-        } catch (Exception ex) {
+            try {
+
+                FileUtils.deleteDirectory(new File(System.getProperty("user.home") + "\\.temevi\\pitreport"));
+
+            }catch(IOException e){
+                System.err.println("Non sono riuscito ad eliminare qualche file contenuto nella cartella pitreport. \n Dettagli: \n"+e.getMessage());
+            }
+
+        } catch (IOException ex) {
             ex.printStackTrace();
         }
